pyproject.toml
[tool.poetry]
name = "agents-course"
version = "0.1.0"
description = ""
authors = ["hench89 <henrik.c@maersk.com>"]
readme = "README.md"
packages = [
  {include = "*", from = "src"},
]

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
smolagents = {extras = ["litellm"], version = "1.9.2"}
sqlalchemy = "^2.0.41"
huggingface-hub = "^0.31.4"
duckduckgo-search = "^8.0.2"
langgraph = "^0.4.5"
langchain-openai = "^0.3.17"
langchain-core = "^0.3.60"
datasets = "^3.6.0"
langchain-community = "^0.3.24"
rank-bm25 = "^0.2.2"
gradio = {extras = ["oauth"], version = "^5.31.0"}
wikipedia = "^1.4.0"
wikipedia-api = "^0.8.1"
youtube-transcript-api = "^1.0.3"
openai-whisper = "^20240930"
pdfplumber = "^0.11.6"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
